Refactored timer and twitter to modules due to no data encapsulation / broke out classes and modules to individual files
